api/segment_anything/requirements.txt
litserve>=0.2.8
numpy>=1.24.0
pillow>=11.2.1
torch>=2.7.0
sam-2 @ git+https://github.com/facebookresearch/segment-anything-2
transformers>=4.51.2
